requirements.txt
fastapi 
uvicorn 
langchain 
langchain-openai 
openai 
chromadb 
python-dotenv